Legacy Product Finder/requirements.txt
Flask==2.2.3
gunicorn==20.1.0
mysql-connector-python==8.0.33
Werkzeug==2.2.2